+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - `\Wizaplace\SDK\Catalog\ProductVideo::getVideoUrl` now returns an `\Psr\Http\Message\UriInterface` instead of a `string`
 - `\Wizaplace\SDK\Image\ImageService::getImageLink` now returns an `\Psr\Http\Message\UriInterface` instead of a `string`
 - Removed `\Wizaplace\SDK\Catalog\Product::getCreationDate`
+- Removed `\Wizaplace\SDK\Catalog\ProductSummary::getCategorySlugs`
+- Removed `\Wizaplace\SDK\Catalog\Product::getCategorySlugs`
 
 ### New features